+++ Proyecto-ODC/MetodoMultiplicacionPE.h
@@ -1,6 +1,6 @@
 #ifndef METODOMULTIPLICACION_H_INCLUDED
 #define METODOMULTIPLICACION_H_INCLUDED
 
-long * parteEnteraOrigenA10 (char * numero , short * baseInicio, short * mostrar);
+long long* parteEnteraOrigenA10 (char * numero , short * baseInicio, short * mostrar);
 
 #endif // METODOMULTIPLICACION_H_INCLUDED
